Send metrics in Bulk, create csv, fixed folder creation issue
Modified script to send metrics in a bulk payload instead of
individually. This drastically increased performance.
In order to do this, needed to modify existing send-bulkgraphitemetrics
logic, so created a send-v2_bulkgraphitemetrics so they don't get
confused.
Script now creates a csv file with the metrics that were sent, in the
same directory as the final nar.
Noticed the folder creation under the nar folder for the array wasn't
working properly; fixed this.
Added all functions required directly into script.